feat: パラメータバリデーション責務分割とカプセル化
- バリデータ責務分割: ZeroParamsValidator, OneParamsValidator, TwoParamsValidator 作成
- params_parser初期分岐でValidatorに精査を任せる実装
- zero,one,twoのインスタンス生成し、バリデーション結果を受け取る構造
- 受け取ったバリデーションの成功したものを判定結果に使う方式
